pcbnew: use doubles to pass line widths in the DXF importer as they are expressed in DXF's units and can case severe rounding errors when passed as ints

+++ pcbnew/import_gfx/dxf_import_plugin.cpp
@@ -1201,7 +1201,7 @@ void DXF_IMPORT_PLUGIN::addPoint( const DL_PointData& aData )
 
 
 void DXF_IMPORT_PLUGIN::insertLine( const VECTOR2D& aSegStart,
-                                    const VECTOR2D& aSegEnd, int aWidth )
+                                    const VECTOR2D& aSegEnd, double aWidth )
 {
     VECTOR2D origin( SCALE_FACTOR( aSegStart.x ), SCALE_FACTOR( aSegStart.y ) );
     VECTOR2D end( SCALE_FACTOR( aSegEnd.x ), SCALE_FACTOR( aSegEnd.y ) );
@@ -1216,7 +1216,7 @@ void DXF_IMPORT_PLUGIN::insertLine( const VECTOR2D& aSegStart,
 
 
 void DXF_IMPORT_PLUGIN::insertArc( const VECTOR2D& aSegStart, const VECTOR2D& aSegEnd,
-                                   double aBulge, int aWidth )
+                                   double aBulge, double aWidth )
 {
     VECTOR2D segment_startpoint( SCALE_FACTOR( aSegStart.x ), SCALE_FACTOR( aSegStart.y ) );
     VECTOR2D segment_endpoint( SCALE_FACTOR( aSegEnd.x ), SCALE_FACTOR( aSegEnd.y ) );
@@ -1297,7 +1297,7 @@ void DXF_IMPORT_PLUGIN::insertArc( const VECTOR2D& aSegStart, const VECTOR2D& aS
 
 #include "tinysplinecpp.h"
 
-void DXF_IMPORT_PLUGIN::insertSpline( int aWidth )
+void DXF_IMPORT_PLUGIN::insertSpline( double aWidth )
 {
     #if 0   // Debug only
     wxLogMessage("spl deg %d kn %d ctr %d fit %d",
